requirements.txt
# Versiones CPU-only (sin soporte para NVIDIA)
torch==2.2.2+cpu
torchvision==0.17.2+cpu
torchaudio==2.2.2+cpu --index-url https://download.pytorch.org/whl/cpu

# YOLO y Streamlit (sin forzar dependencias GPU)
ultralytics==8.2.0
streamlit==1.34.0

# Procesamiento de imágenes (ligero)
opencv-python-headless==4.9.0.80
pillow==10.3.0
numpy==1.26.4